сгенерирован файл init-schema.sql на основе сущностей из папки models, этот файл разворачивает бд

+++ src/main/java/ru/databasePetProject/RestAppUniversityProject/models/CultureHouse.java
@@ -19,7 +19,7 @@ public class CultureHouse {
     @Column(name = "id_culture_house")
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private short idCultureHouse;
+    private long idCultureHouse;
     @NotNull
     @Column(name = "address")
     private String address;
